move custom reqs tests
