+++ setup.cfg
@@ -23,7 +23,6 @@ markers =
     long: Marks long-running tests
 testpaths =
     tests
-    tests_requirements
 addopts = -rav --durations=0 --cov=mlem --cov-report=term-missing --cov-report=xml
 
 [mypy]
