+++ tests/core/custom_requirements/pkg/subpkg/__init__.py
