+++ tests/core/custom_requirements/pkg_import.py
