+++ tests/core/custom_requirements/proxy_pkg_import.py
