+++ tests/core/custom_requirements/use_model_meta.py
